mojo: Reorganize all Mojolicious app state under single namespace.
In a move to further the goal of making it at least possible for a
single running kdesrc-build web server to manage multiple build contexts
(as identified by the kdesrc-buildrc path), continue the work to move
per-request state for the web server in an appropriate spot.

In this case, I migrate multiple Mojolicious app fields related to a
build under the new ksb_state (named this way to namespace the shared
dictionary against possible Mojolicious uses).

What remains is to migrate the package-scoped variables to also live
under this object. This would permit future separation of this data into
per-build-context objects all tied to a single web server.