mojo: Move metadata loading to after server launch.
This change seems simple enough but is (or should be) a big enhancement
to the developer experience, especially if the sysadmin/repo-management
metadata hasn't been pulled recently.

Unfortunately we can't proceed to looking for modules to build until the
metadat is available so I also update the U/I to check for the metadata
and spit out a message after 30 seconds if it's still not loaded since
then something probably *is* wrong.