Return default folder structure to flat
The "invent" structure results in fewer folders in the src folder, but
requires familiarity with the invent.kde.org structure to be able to
find anything, making it a worse default especially for new
contributors. For example, with "flat", to find dolphin, you just go to
the "dolphin" folder. With "invent", you have to know that Dolphin lives
in the "system" group and go to "system/dolphin." If you don't know
this, you have to hunt for it in every group. Quick navigation requires
internalizing the parent group for every single KDE repo, rather than
just having to know the repo's name.

Similarly, to quickly move from one repo to another, you have to
remember an additional piece of information compared to using the "flat"
layout. For example if you are in "kio" and wand to go to "dolphin", you
have to do `cd ../../system/dolphin` rather than just `cd ../dolphin`.
It is similarly cumbersome if you use a GUI file manager rather than the
command line.

For these reasons, I believe "flat" is the better default while "invent"
is more suited for experienced contributors, and as such they should
turn it on manually.