buildsystem: Apply the setting of num-cores by default.
This change turns what was previously a 2-part process to pass the
setting in num-cores into the build command, into a single step.

Previously you had to have the user set num-cores and then separately
use something like make-options, ninja-options, or "set-env MAKEFLAGS"
to pass that to the build command.

This was not done earlier, partially because I expected it to be
complicated to fix the various build systems to do something appropriate
with this setting.

But it turns out I'd already centralized the build process through a
slightly generic combination of ksb::BuildSystem::buildInternal and
ksb::BuildSystem::safe_make (which despite the name can run any
specified command as a build command).

Since all supported build systems use the -j flag to control the number
of build jobs, I can just adjust the build flags here to support all
builds. So I do.

This only controls the flags passed to the build command directly, so
indirect methods like "set-env NINJAFLAGS" are still needed to support
use cases like Qt5WebEngine. But ultimately I would like to remove
num-cores-low-mem entirely and just let users change the setting for
num-cores on a per module basis, as with any other option.

This change obviates the need for !105 and fixes #75.