Fix broken test cases with new max cores logic.
I agreed with the change to take off a CPU but thought that would apply
to the auto-detection logic (I feel a user should be able to use all
their CPUs if they specifically ask for it and their system can
otherwise handle it).

Since this logic impacted a recent test I opted to refactor it slightly
(so that it no longer shaves off a CPU unless you've asked for more than
you have, in which case we assume a broken config and use the same logic
as if we'd auto-detected it).

I still end up needing to improve the test case though, in case CI runs
this on a system where nproc would return 1, 2 or 3.