mojo: Don't send events to broken WebSocket connections.
This causes an exception that will stop the whole build process and is
possible in --backend mode, where closing a web browser tab during the
pointed to http://localhost:$PORT/event_viewer during the build will
cause the WebSocket to close on the kdesrc-build but we were still
pushing events to it.

Fixes issue #66.