Fix lockfile handling and avoid interactive request for stale lockfile.
This ended up being somewhat involved to ensure that the execution
sequence always proceeds in a logical order, even if a stale lockfile is
overwritten or (harder case) if we should abort the build from within a
bunch of promises.

This required moving up some of the calls that could cause the build to
have to fail.

In the process I also fixed some errors to be more consistently
displayed should they occur, and for the Mojolicious logs to continue to
write to mojo-backend.log even once the backend starts building.