Update sample-rc to reflect changes to defaults
Defaults to settings and comments have changed but are not being
reflected in the `sample-rc` which populates a user's
`kdesrc-buildrc` file when running `--initial-setup`.

This is especially important as potential new contributors
are instructed to run `./kdesrc-build --initial-setup`, and this
leaves them in a less than ideal state; for example the
`compile_commands.json` will not be populated by default,
therefore there will be no LSP support as expected with
https://invent.kde.org/sdk/kdesrc-build/-/merge_requests/100 .

This change brings the inital config more in line with
the defaults generated when running `kdesrc-build-setup`.