mojo: Only force-reject build promise if build would continue.
This addresses a spurious Mojolicious warning about an unhandled
rejected promise after a dependency phase failed to complete, like
this example for kguiaddons when a ki18n failure with --stop-on-failure
kills the build:

    Unhandled rejected promise: Prerequisite to kguiaddons/build failed
    at /opt/kdesrc-build/modules/Mojo/Reactor/Poll.pm line 129.

The issue is that ki18n's failure during --stop-on-failure caused any
still-waiting promises to reject before they run, including promises
that do not yet have anything directly awaiting its result with a
catch(...) handler.

This is unneeded when --stop-on-failure is true because the whole build
will stop anyways. So we prevent the warning by only rejecting the
promise if the build would continue, so that dependent items phases do
not run.