mojo: Fix one bug with --no-src build.
The check for whether an update job has been queued (used to add an
ordering dependency on the build job for the same module in this
situation), actually creates an item entry for the update in the data
structure that eventually turns into the promise chain.

Replace this with a method that doesn't create the item if it doesn't
already exist, to make other checks do their job properly.

--no-src is still broken though but one bug down...