Re-add strict/warnings for kdesrc-build pre-setup.
We can't 'use ksb' or any other module in modules/ until after the
FindLib stuff has done its magic.  So manually enable strict/warnings in
initially before we do that.

I also noticed the error messages intended for bare/containerized
distros are not super-helpful when they happen due to Perl syntax errors
and the like, so get rid of error-$$.log and just output the error
message to the console.