Enable Perl subroutine signature for all modules.
This lets us declare Perl subroutines with a syntax closer to modern
languages.

Unfortunately it's not directly compatible with Perl subroutine
prototypes (which affect how Perl syntactically treats subroutines).

While diagnosing this I decided it would be a good idea to have a test
case which simply loads every ksb module. I scripted this initially but
it would need to be maintained manually as new modules are added or
removed.