buildsystem: Cleanup safe_make, start on removing stale subdir feature.
The 'subdir' stuff complicates the code quite excessively and is almost
certainly broken. It's a holdover from KDE 3 era when the CVS and SVN
modules were a large aggregation of multiple libs and applications that
could be built independently (using the 'checkout-only' option to filter
the subdirs being built down).