boilerplate: Relax Perl version requirement and avoid new modules.
We don't need the 'experimental' module (it's nice and it's core but we
can easily achieve the effect we need to avoid adding dependencies).

Plus (although this is fixed now) we're not using anything in Perl 5.28
and are not likely to in the future, and some CI might still be using
Perl 5.26 like ours was.

See also !126 which is now unneeded.