first-run: Avoid dependence on ksb::Util for minimial distros.
Many Linux distros ship base images for Docker which have extremely
minimal Perl installs (if they come with Perl at all).

ksb::FirstRun (which underpins kdesrc-build --initial-setup) is meant to
survive in this condition (as long as Perl is installed). But relying on
modules from ksb::Util breaks this on Debian (which doesn't include
File::Find in its base Perl).

Since ksb::FirstRun is the only user of the ksb::Util function it was
calling, just move the function over. I also added some comments to
remind me next time that this needs to be something we check.

An easy way to check on Docker is with a command line:

    docker run --rm -it -v /path/to/kdesrc-build:/opt/kdesrc-build:ro \
        debian:latest sh -c "/opt/kdesrc-build --initial-setup"

As of now this still has problems, but at least the script will load and
then note the installer failed rather than crashing with a Perl
exception.