Do not set custom CMake parameters for KActivities
This causes the default CMake options to be overridden and in my case
the KF6 KActvities build used Qt5, because the -DBUILD_WITH_QT6=ON was not used.

We actually do not need the Cmake options at all, because BUILD_TESTING is handled by ECM
and the check if exceptions are needed is done as part of KActivities.