buildsystem: Return the build results object from buildInternal
Previously we just returned a boolean flag that throws out the rest of
the results object. Of course the object itself had only a bool but this
enables us to return more metadata and make decisions based on that,
such as whether the build did any work (which can influence a decision
to install or not).