build: Avoid install phase if we know the build phase did nothing.
When a Ninja-based generator is in use, we can look for its "ninja: no
work to do" message to know that the build phase did nothing.

In this situation, there's no point to installing, so why bother.

This has an additional improvement as well, we don't bother *uninstall*
when use-clean-install is set just to reinstall the same files.

This is at best a heuristic as it's possible that a changed dependency
library that is not part of the source repo could influence the binary
that is built.  In this situation ninja would report no work to do even
though the binary would change.  So I also tell the user how to force an
install (by forcing the build to do work using --refresh-build, other
methods might work as well).

Fixes #92.