build: Ensure a rebuild following an interrupted build is attempted.
Previously, interrupting a module build during the build was not treated
as a build failure, so the module's persistent options were not changed.
In the normal scenario this means the module's last build was remembered
as a success. Then by default, when you'd go to rebuild the module right
after, the build would be skipped because the source didn't change.

BUG:289754