kdesrc-build-setup: Make ksb:: modules available to this script.
This will make it possible to start sharing code between
kdesrc-build-setup and the equivalents in kdesrc-build --initial-setup
(which uses ksb::FirstRun).  Of interest to #38, tracking that the
separate ways to do this same thing should merge.

Tested on my system and also by running an Alpine container (for both
in-repo and from-installed modes), though there's a separate bug in
Alpine I'll fix in a bit.