mojo: Simplify UI / backend comms.
This makes the commit from a few months back actually work with
--no-src.

It seems that I was having difficulty getting all the possible scenarios
involving syncing up two different event streams from backend to client
always working 100%.

So although it requires polling on the client's part, I've shifted away
from a WebSocket long-polled connection for now.

This is mostly for completionism's sake, as I intend to shift my effort
back to the 'master' branch from here, though that is likely to include
Mojolicious (though maybe not trying to take advantage of all of its
power this time).