git: Don't fail to clone due to branch name if none required.
There is a wider variety of potential default branch names from a git
repository that we might want to clone.  Many use now 'main' but not all
of them. We also have modules that use something else entirely, like
'dev' (as reported in bug 436739).

Most of the time the user doesn't actually care about the branch, they
just want whatever the mainline trunk is. To handle this we had
previously defaulted to assuming the branch was 'master' so we could
optimize the amount of data transferred during initial clone.

Now this is likely to lead to an error. Instead, if the user hasn't
specified otherwise, just perform the git clone as normal.

To do this right there's a wee bit of other work elsewhere to detect
which branch head we actually got from the remote git repository but
that's not difficult to do either.

BUG:436739