util: Simplify inspect-during-test function.
For a previous unit test I found it useful to allow a function to be
called from the normal code but try to override the function. The way I
implemented it is very clunky though, with log_command calling a
function which is normally AUTOLOAD'd into being a no-op, but is defined
during the unit test to extract the data it needs.

This clunky implementation was because I couldn't figure out how to
override the ksb::Util::log_command sub directly.  With Mojo's
well-tested code available I though to use Mojo::Util's monkey_patch.

But it still didn't work.

It took longer than I'd like to admit but I figured out that the reason
none of these attempted worked up until now is because I was overridding
the wrong package's "log_command" entry.

In short, when ksb::BuildSystem::KDECMake is loaded, it ultimately
resolves the bare 'log_command' to an entry in *its* own symbol table
resolving to the subroutine in ksb::Util.

You would think this wouldn't be a problem because it should copy
ksb::Util's overridden log_command. But Perl modules loaded through
'use' happen in a compile phase before normal code execution, and so the
'use ksb::Application' in the test, even if it was the last line, was
enough to cause ksb::Module and then ksb::BuildSystem::KDECMake to be
loaded, before our override would be installed.

There are two possible fixes:

1. Override the 'log_command' in ksb::BuildSystem::KDECMake afterwards,
   as this is the one we're trying to introspect, or
2. Use a BEGIN { } block to force the monkey patch to happen before
   ksb::Application ends up loading the rest of the modules.

I chose #2 for this one as it feels more correct to ensure all modules
using ksb::Util get the same version of log_command.

This will permit future work to move log_command to use
Mojo::IOLoop::subprocess to do its work instead, since the old code for
this is now dead and removed.