git: Use a more reliable check for remote default head.
The previous check I had, using git symbolic-ref, doesn't seem to work
after initial clone if you're since switched to a detached HEAD (as with
a tag).

"git remote show" has output we can scrape to get the information we
need in this situation, but it is astonishingly slow to do it this way
as there's a whole bunch of other metadata git remote show will give you
that we don't need for this.

What I've settled on is reading the appropriate symbolic ref file in
.git/refs/remotes/$remoteName/HEAD, which points to the right branch in
all the testing I've been able to do. I'm sure something will break it
later though.