ui: Show all log directories used in the build if there are multiple.
It is possible for there to be multiple log-dir settings in use in a
single build.  If this happened only the global log-dir setting would be
shown in the ending message about where logs for the run were put.

This can be especially confusing if you're only building one module and
it's *not* using the global log-dir setting.

To fix this, look for "base log-dir" values that were actually used
during the build, and cleanup log handling so that we don't call
getLogDir unless we're actually about to fork child processes (which was
ultimately why ModuleResolver had been forcing a call to getLogDir, I
moved this to TaskManager instead).