autotools: Convert to use promises.
This one is especially interesting because we have a series of
log_command calls that we were able to turn into a chain of promises,
including use of rejecting promises to allow early exit from that chain.

Tested with nasm, which doesn't actually build due to other problems,
but manages to run autogen.sh and then configure (it's getting hard to
find autotools-using codes, luckily for us...).

Once configureInternal is converted everywhere to return promises we'll
be able to simplify handling of the update/build multiprocessing code.