ipc: Do not print messages to user from IPC handler.
The message to the user about why we were proceeding to build a module
even though no source code changes had occurred was being printed by the
IPC handler, rather than the handle_updates() or a related function.

This was intended to support having the output pop up at the right time,
when the buildSingleModule function called ipc->waitForModule().

However, the IPC handler that emits the message via waitForModule() is
also called through a separate catch-all handler that happens when the
IPC handler is being torn down.

This causes module update spam to be sent to console if ever the build
thread is aborted early.  Instead let the message be shown to the user
only if we're doing updates so that IPC teardown can be silent like it
should be.