buildsystem: Be more restrictive on when to skip install.
This fixes a bug where the documentation of how to force kdesrc-build to
proceed with the install "--refresh-build this module to force install"
did not match kdesrc-build's behavior.

But in addition this adds a check that the module was ever successfully
installed by checking for a matching persistent option.  I suspect this
might be part of the problem experienced by a user in bug 456593.

CCBUG:456593