build: Port runBuildCommand to use a promise instead of log_command.
This one was difficult because runBuildCommand uses the 'child output
callback' feature of log_command heavily to do what it does, looking for
progress output and detecting warnings.

Making this work with a promise-returning logging command would make
every such function call quite involved if we just did this by adding
more arguments or an 'options' block as done with log_command.

Because of this I opted to use a 'builder' interface for this, which
required a new package (ksb::Util::LoggedSubprocess)

Note that since we have to capture the child's output we're already
forced to fork another child process, so doing all this from under
Mojo::IOLoop::Subprocess might be overkill. Instead we might be able to
monitor for the child output's filehandle from within the parent
directly using Mojo::Reactor or similar. However I understand what's
going on here so I'm going to go with this for now.

Note there are two ways to handle capturing and processing the child
output: 1) do the processing/filtering in a callback within the child
and then have the child feed interesting bits back to the parent using
->emit(progress => ...), or 2) send all child output to the parent using
the same progress event, and then process on the parent's side.

This implements 2) for now, again mostly for expediency.  But it seems
plenty fast and progress reporting for runBuildCommand in particular is
going to require frequent updates either way.