IPC: Print messages from update thread as we receive them if waiting.
This is as opposed to waiting until the module's entire update has
completed and then printing all the module's update messages as once.

This allows for commands that need git-clone to run as the first module
to at least print the message that the clone is in progress even while
the git-clone is running.

Note that we still wait to print update messages until after we've
started waiting for the module.