async: Migrate most existing promise->wait to error-checking helper.
The error checking is nice but it's important not to blindly do a
search-and-replace here, as the test suite caught an error in one of the
replacements that ended up not confirming to the containing function's
return style.

Still, important to get this added early as it only gets harder to add
the longer I wait, and these functions are important to avoid a class of
error that is possible if you try to call promise->wait re-entrantly.
Turns out Qt is not the only library that doesn't like nested event
loops.