Actually default to stopping on failure
This is clearly the intention in the code, seeing as how
"stop-on-failure true" is written into the default config files in two
places. But if for some reason your config file manages to not have that
in it, you'll fall back to the internal global default, which was still
false. Let's change that to true.