Remove Net::DBus dependency.
We use Net::DBus for the power-profiles-daemon. Net::DBus is a
comprehensive module intended to support the breadth of everything about
D-Bus. This makes it a heavyweight dependency by kdesrc-build standards,
and indeed some distros don't even have packages for it (forcing uses to
go through something like perlbrew or App::cpanminus instead).

However all we need to do with D-Bus is to make a D-Bus connection (the
life of the connection is what determines the lifetime of the profile
hold we place with power-profiles-daemon, or we could just use
dbus-send), and a single D-Bus call.

While I wouldn't call this completely straightforward, it's not
difficult to build support for just enough of the D-Bus protocol to do
this ourselves, since we have already taken Mojolicious as a dependency.

So this change reverts the Net::DBus dependency without getting rid of
the feature we used it for.

The test suite passes, simple Alpine-based testing with 'kdesrc-build
--initial-setup' is happy again, and running dbus-monitor --system
confirms that the changed code still successfully connects to the D-Bus
and makes the required call.