Remove ffmpeg-devel and vlc-devel from Fedora first-run package list
These packages require the RPMFusion repo to be installed, wihich is a
manual action that must be done and we can't ensure that all
kdesrc-build users have done it. Remove these for now.

This will prevent building phonon-backend-vlc successfully until
RPMFusion is added and those packages installed, but that's
unfortunately unavoidable on Fedora right now.