ui: Don't spam about updated modules if --stop-on-failure activates.
MR !178 gives an example of poor output from kdesrc-build if a module
fails to build and causes kdesrc-build to stop early.

But rather than try to clean up the message, we should probably not list
the modules at all.  In this situation, with update thread finishing
after the build thread (rather than before as normal), failures are
still listed to the user and we don't need to announce success.

So just make this message a simpler notice instead, and make it display
only in verbose mode.