Do not suppress deps if a partial selector is mentioned on command line.
The issue is that mentioning a 'partial' selector (a selector based on
a portion of the project hierarchy, like "kde/workspace") on the command
line could cause the `include-dependencies` config option to be
disabled.

So a command like "kdesrc-build plasma-desktop" (with
include-dependencies true in the kdesrc-buildrc) would find all
KDE-based dependencies and include them in the build.

But a command like "kdesrc-build workspace", or even "kdesrc-build
plasma-desktop workspace", would build *only* the modules under
kde/workspace, no matter what you set include-dependencies to.

This is due to an improper check for the presence of the
--no-include-dependencies command line option.  The current check looks
for --include-dependencies and assumes that if it is not present, that
there was implicitly a --no-include-dependencies on the command line and
disables include-dependencies on the ksb::Modules the resolver generates
as it expands the selector in a module list.

I fix this for now by defaulting the command-line value to the current
config setting if nothing is set on the command line.

Fixes #115.