Git.pm: Use --includes when querying insteadOf/pushInsteadOf
The user may have already added KDE's URLs to a separate file and then
included that file in ~/.gitconfig. Unlike regular git commands,
`git config --global` does not process includes without `--includes`.

This commit does not change the behavior for adding new options. If the
config options are missing, they are still added to the main
~/.gitconfig.

Signed-off-by: Andrew Gunnerson <accounts+kde@chiller3.com>