dep-resolv: Add test code for reading V2 format dependencies.
Keeping up with the push to get rid of `projectpath` in
sysadmin/repo-metadata [1], kdesrc-build needs to migrate to a different
dependency graph format.

Since this is already being generated instead of manually-maintained,
there's no reason not to make this a JSON blob while we're at it.  Using
the sample JSON data based on repo dependencies [2], this change allows
kdesrc-build to read in the dependencies encoded in that JSON file and
still permit --include-dependencies to work in my limited testing.

Right now it is guarded by a specific env. var that must be set,
otherwise the existing code is used instead.  Once the format is stable
we can shift users over, and then eventually remove the old format data.

[1]: https://invent.kde.org/sysadmin/repo-metadata/-/issues/2#note_602604
[2]: https://invent.kde.org/-/snippets/2497