'fix' path to kf6-qt6-build-include in kdesrc-buildrc-kf6-sample
The path listed here only makes sense if kdesrc-build was installed via
"kdesrc-build kdesrc-build".

However, that isn't documented in
https://community.kde.org/Get_Involved/development#Set_up_kdesrc-build,
and seems unnecessary in general.

The expectation seems to be that the user manually fixes that to match
the actual install location, so make that more obvious

This also matches the kf5-qt5 sample