Associate build data with manually-specified rc file
Currently, the build data is stored globally when using the default
rcfile, and when using a manually-specified rcfile, it is stored in
the same directory as the rcfile. This generally works.

But if you have two rcfiles that live in the same directory, they
inappropriately share build data, because each one stomps on the build
data file generated by the other one. You can manually specify
`persistent-data-file` in your config file to fix this, but that may be
forgotten, not known about, or unwanted.

This commit fixes that by appending the name of the manually-specified
rcfile to the build data file when not using `persistent-data-file` in
the config file. This automatically associates each manually-specified
rcfile with its own private store of build data, even when it lives in
the same directory as another rcfile.

There is no change when using the default global rcfile behavior.

Fixes #117