Accept also linked worktree directories
Directories that contain a linked worktree don't have a `.git`
folder, but a `.git` file which contains the route to the
worktree info in the original repository.

Git commands do still work as usual on this directories, so
just expanding the check for a .git` file makes kdesrc-build
work on a git worktree link of the repo.