Fix openSUSE missing package version
In the most recent Tumbleweed version, `ruby3.1-rubygem-sass` is no
longer available in the package manager; the available version is
`ruby3.2-rubygem-sass`.

This means we are breaking compatibility with openSUSE Leap which still
uses the old version, but devs are more likely to be using Tumbleweed
than Leap, so this is an acceptable compromise for now.