Set rpath to kdedir for appstream, which uses meson
Without this, appstreamcli and libAppStreamQt.so (used all over
Plasma, including Discover) would rely on LD_LIBRARY_PATH to find
the correct libappstream.so. That's not great because the standard
prefix.sh doesn't set LD_LIBRARY_PATH, relying on rpath instead.

Appstream's meson build doesn't set rpath for installed binaries
by default, but can be prodded into adding it with LDFLAGS.

In order to provide the correct name of the library directory
to kdesrc-build, e.g. "lib" or "lib64", this commit introduces
a new option called libname. Its default value gets initialized
in the same way as an earlier shell implementation in
sample-kde-env-master.sh, which doesn't appear to be widely used
at this point but clearly had some experience going into it.