Remove wrong EXCLUDE_DEPRECATED_BEFORE_AND_AT flag from non-KF6 modules
EXCLUDE_DEPRECATED_BEFORE_AND_AT is not a flag to control which deprecated
API of other libraies to use or not.

It is a flag valid for the repo it is used on, to control if any deprecated
API and its implementation of that very repo should be excluded in the
build, and up to which version.

So it also needs to use a version matching the repo.

Currently it is only needed with KF6 for historic reasons. Ideally repos
should set a useful default internally, but that was missed for KF modules.