Add packagekit to dependencies for initial setup
packagekit is required to build packagekit-qt

@teams/qa

I don't know if there's any order in which the packages are listed, so I just put packagekit close to the QT stuff.
I've googled for the package name in Debian, OpenSUSE and Fedora, but have only actually tested that this is correct under Arch. 

BUG: 472850