resolver: Allow 'options' to apply to module-sets.
This change allows for 'options' blocks to apply to entire module-sets.
As with normal 'options' blocks, these don't add the new modules to the
build or update list, so these can be used to provide default options
that can still be overridden later by the user, though this currently
work only if a later 'options' block is module-specific.  See the
updated test case for a deeper example.

Use by adding the 'repository kde-projects' and 'use-modules' entries as
appropriate, example:

    options kf6-frameworks
        repository kde-projects
        use-modules frameworks

        cmake-options -DBUILD_QT6=ON -DBUILD_QT5=OFF
    end options

Note that unlike with module/module-sets, there is no separate declaration
keyword.