resolver: Allow "options" to be overridden by later module-sets.
This builds on the last commit to allow module-sets that are defined
*after* an options block (which now can contain options for an entire
module-set) to override options defined in the previous 'options' block.

This means an 'options' block can be defined in, say, a sample
configuration file without forcing the user to actually build the module
set.  But if the user loads it into their own config the pre-defined
options should apply without breaking the user's own customizations.