Don't build all of extragear and playground by default
This `module-set` block was added to default BUILD_WITH_QT6 to
ON for any of the specified module sub-tree in kde-projects.

However, using `module-set` ends up pulling all of the modules
in those sub-trees into the active build set, building many
dozens of modules that cannot actually be built with Qt/K6 yet.

Commit fa11f2ee recently added a better way of doing this to
kdesrc-build, which is to just add an `options` block. This
is also moved up here, allowing the files included after to
further override the options in other blocks.

Manually rebased from Eike's submission in !280