util: Remove read-only files under build dir as well.
A user reports that kdesrc-build was unable to remove the build
directory of zxing-cpp, and managed to narrow it to a git directory for
some of its dependencies being created in the build directory.

Some of the git files are set to be read-only, this causes File::Path's
`remove_tree` function to fail as it will not remove read-only files
with `safe` turned on.

This is the only documented effect of `safe` as this is the exact
scenario it is meant to prevent. However, kdesrc-build has always owned
the contents of the build directory, so `safe` is not helpful or needed
here.

To fix the issue, remove the `safe` usage and add a test case to prevent
future recurrence. The test case fails without the change and passes
with it.