Add support to generate architectural info from
Uses kde's Codevis to generate a project containing
the code architecture information from all kde projects
build by kdesrc-build, if there's a `compile_commands.json`
on the build folder.

This is *really* time consuming, and I have tried to add
an option to disable it, but my perl is weak and I didn't
found out how.