Turn on --include-dependencies by default for real
This gets set to true in FirstRun.pm, but if you don't run that and
instead grab a config file from somewhere else (e.g. the KDE Community
wiki), then it won't have "include-dependencies true" set in it and
kdesrc-build will fall back to its internal default of false.

The documentation says it's on by default, and the first run module
turns it on explicitly, so it's clear what the intention is here. Let's
change the internal fallback default to true.