Add "generate-vscode-project-config" option
This change adds `generate-vscode-project-config` option for `kdesrc-buildrc`
and the corresponding command line flag.

When set to `true`, kdesrc-build will create the `.vscode` folder
in the module source directory with configuration that makes VSCode to work
properly with KDE projects, such as setting the correct build
directory & enabling support for C++, CMake, LSP & IntelliSense.

The process specifically checks for an existing `.vscode` folder,
in case the user already has custom settings present - and only
proceeds if it does not exist yet.

Closes #84